+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Ruby client
 - Action input parameters can be given either as a hash using a positional
   argument, or keyword arguments, but not both
+- `HaveAPI::Response#wait_for_completion` now accepts keyword arguments
 
 ## PHP client
 - Requires PHP >= 7.4